pyproject.toml
[tool.poetry]
name = "MicroscopeController"
version = "0.1.0"
description = ""
authors = ["Vladimir Usatikov <vladimir@usatikov.com>"]

[tool.poetry.dependencies]
python = "^3.7"
PyQt5 = "^5.15.0"
logscolor = {git = "https://github.com/v-usatikov/logscolor.git"}
graphic-ext = {git = "https://github.com/v-usatikov/graphic_ext.git"}
motor-controller = {git = "https://github.com/v-usatikov/motor_controller.git"}
numpy = "^1.19.0"
opencv-python = "^4.2.0"
Pillow = "^7.1.2"
pymba = "^0.3.7"
lmfit = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
